requirements.txt
# AI Code Explainer — dependencies

# Backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain + OpenAI
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.30.0

# Frontend
streamlit>=1.35.0
requests>=2.31.0
